dbx-apps-deployment-demo/pyproject.toml
[project]
name = "dbx-apps-deployment-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.13"
dependencies = [
    "databricks-sdk==0.59.0",
    "databricks-sql-connector==4.0.5",
    "python-dotenv==1.1.1",
    "streamlit==1.47.0",
]
